+++ services/device-registry/src/main/java/org/eclipse/hono/deviceregistry/FileBasedTenantService.java
@@ -100,7 +100,7 @@ public final class FileBasedTenantService extends CompleteBaseTenantService<File
             return Future.succeededFuture();
         } else {
             final Future<Buffer> readResult = Future.future();
-            vertx.fileSystem().readFile(getConfig().getFilename(), readResult.completer());
+            vertx.fileSystem().readFile(getConfig().getFilename(), readResult);
             return readResult.compose(buffer -> {
                 return addAll(buffer);
             }).recover(t -> {
@@ -118,7 +118,7 @@ public final class FileBasedTenantService extends CompleteBaseTenantService<File
         } else if (vertx.fileSystem().existsBlocking(getConfig().getFilename())) {
             result.complete();
         } else if (createIfMissing) {
-            vertx.fileSystem().createFile(getConfig().getFilename(), result.completer());
+            vertx.fileSystem().createFile(getConfig().getFilename(), result);
         } else {
             log.debug("no such file [{}]", getConfig().getFilename());
             result.complete();
@@ -187,7 +187,7 @@ public final class FileBasedTenantService extends CompleteBaseTenantService<File
 
                 final Future<Void> writeHandler = Future.future();
                 vertx.fileSystem().writeFile(getConfig().getFilename(),
-                        Buffer.factory.buffer(tenantsJson.encodePrettily()), writeHandler.completer());
+                        Buffer.factory.buffer(tenantsJson.encodePrettily()), writeHandler);
                 return writeHandler.map(ok -> {
                     dirty = false;
                     log.trace("successfully wrote {} tenants to file {}", tenantsJson.size(),